refactor: flatten util
