+++ src/cala/nodes/iter/detect/catalog.py
@@ -6,7 +6,7 @@ from xarray import Coordinates
 
 from cala.assets import Footprint, Footprints, Movie, Trace, Traces
 from cala.models import AXIS
-from cala.util.new import create_id
+from cala.util import create_id
 
 
 class Cataloger(Node):
