+++ tests/test_util/test_frame.py
@@ -3,7 +3,7 @@ from datetime import datetime
 import numpy as np
 
 from cala.assets import Frame
-from cala.util.new import package_frame
+from cala.util import package_frame
 
 
 def test_package_frame():
